+++ packages/core/src/functional-components/tooltip/tooltip.interfaces.ts
@@ -12,4 +12,5 @@ export type TooltipOptions = {
   topPositionSmallViewPort?: boolean;
   halfMainAxisOffset?: boolean;
   forceVisible?: boolean;
+  enableInnerScroll?: boolean;
 };