+++ storybook/src/components/info-button/info-button.css-template.ts
@@ -1,11 +1,11 @@
 import { InfoButton } from "dso-toolkit";
-import { html } from "lit-html";
+import { TemplateResult, html } from "lit-html";
 import { classMap } from "lit-html/directives/class-map.js";
 import { ifDefined } from "lit-html/directives/if-defined.js";
 
 import { ComponentImplementation } from "../../templates";
 
-export const cssInfoButton: ComponentImplementation<InfoButton> = {
+export const cssInfoButton: ComponentImplementation<InfoButton<TemplateResult>> = {
   component: "infoButton",
   implementation: "html-css",
   template: () =>