+++ CHANGELOG.md
@@ -6,6 +6,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) a
 ## Next
 
 ### Added
+* Info Button: Toggletip functionaliteit ([#3395](https://github.com/dso-toolkit/dso-toolkit/issues/3395))
 * Onboarding Tip: WCAG verbeteringen ([#3301](https://github.com/dso-toolkit/dso-toolkit/issues/3301))
 
 ### Changed